refactor: split FPGA base and board classes
